requirements.txt
django==1.9.2
graphene[django]
django-graphiql
django-oauth-toolkit
django-cors-headers
django-extensions
django-filter
django-debug-toolbar
django-debug-panel
python-social-auth
requests-oauthlib
psycopg2
jinja2
sendgrid
uwsgi